Refactored HelmReleaseVersion as inner class of HelmRelease.
